Rename change password test file and relocate templates
