+++ users/templates/emails/password_reset_email.html
